requirements.txt
psycopg2==2.9.9
bcrypt==4.1.3
python-dotenv==1.0.1